# design/swagger.yaml
swagger: "2.0"
info:
  version: 0.0.1
  title: Postcard API
  description: Postcard REST API
  contact:
    name: Kari Kangas
    email: karijkangas@gmail.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
host: localhost:4000
basePath: /v1
schemes:
  - http
...
paths:
...
  /registrations/available:
    get:
      summary: Check registration availability
      description: >
        Check registration availability. Currently, only email check is supported.
      produces:
        - application/json
      parameters:
        - name: email
          in: query
          type: string
          description: Email
      responses:
        200:
          description: Success
          schema:
            $ref: '#/definitions/RegistrationAvailability'
        400:
          description: Invalid data
          schema:
            $ref: '#/definitions/Error'
      tags:
        - requests
  /registrations:
    post:
      summary: Submit registration request
      description: >
        Submit registration request. Service contacts the user via email with instructions how to complete the registration process. On success, returns request expiration time as numbers of milliseconds elapsed since January 1, 1970 00:00:00 UTC. Returns an error if the input data is invalid or the user (email) already exists.
      consumes:
        - application/json      
      produces:
        - application/json
      parameters:
        - name: data
          in: body
          schema:
            $ref: '#/definitions/RegistrationRequest'
      responses:
        202:
          description: Success
          schema:
            $ref: '#/definitions/ExpiringRequestResponse'
        400:
          description: Invalid data 
          schema:
            $ref: '#/definitions/Error'
        409:
          description: User already exists
          schema:
            $ref: '#/definitions/Error'
      tags:
        - requests
  /registrations/{id}:
    post:
      summary: Complete registration request
      description: >
        Completes and removes the registration request. The request is also removed if the request fails with an error 409 User already exists.
      consumes:
        - application/json      
      produces:
        - application/json
      parameters:
        - in: path
          name: id
          required: true
          type: string
          description: Registration request id
      responses:
        201:
          description: Success
          schema:
            $ref: '#/definitions/CompleteRegistrationResponse'
        404:
          description: Not found 
          schema:
            $ref: '#/definitions/Error'
        409:
          description: User already exists 
          schema:
            $ref: '#/definitions/Error'
      tags:
        - requests
  /password_resets:
    post:
      summary: Submit password reset request
      description: >
        Submit password reset request. Service contacts the user via email with instructions how to complete the reset process. On success, returns request expiration time as numbers of milliseconds elapsed since January 1, 1970 00:00:00 UTC.
      consumes:
        - application/json      
      produces:
        - application/json
      parameters:
        - name: data
          in: body
          schema:
            $ref: '#/definitions/PasswordResetRequest'
      responses:
        202:
          description: Success
          schema:
            $ref: '#/definitions/ExpiringRequestResponse'
        400:
          description: Invalid data 
          schema:
            $ref: '#/definitions/Error'
        404:
          description: Not found
          schema:
            $ref: '#/definitions/Error'
      tags:
        - requests
  /password_resets/{id}:
    post:
      summary: Complete password reset request
      description: >
        Completes and removes the password reset request. The request is also removed if the reset fails with an error other than 400 Invalid data, indicating that the new password is invalid.
      consumes:
        - application/json      
      produces:
        - application/json
      parameters:
        - in: path
          name: id
          required: true
          type: string
          description: Password request id
        - name: data
          in: body
          schema:
            $ref: '#/definitions/CompletePasswordReset'
      responses:
        200:
          description: Success
          schema:
            $ref: '#/definitions/CompletePasswordResetResponse'
        400:
          description: Invalid data 
          schema:
            $ref: '#/definitions/Error'
        404:
          description: Not found 
          schema:
            $ref: '#/definitions/Error'
        409:
          description: User not found 
          schema:
            $ref: '#/definitions/Error'
      tags:
        - requests
  /email_changes/{id}:
    post:
      summary: Complete user email change
      description: >
        Complete user email change.
      consumes:
        - application/json      
      produces:
        - application/json
      parameters:
        - in: path
          name: id
          required: true
          type: string
          description: Email change request id
      responses:
        200:
          description: Success
          schema:
            $ref: '#/definitions/CompleteEmailChangeResponse'
        400:
          description: Invalid data 
          schema:
            $ref: '#/definitions/Error'
        404:
          description: Not found 
          schema:
            $ref: '#/definitions/Error'
        409:
          description: User not found 
          schema:
            $ref: '#/definitions/Error'
      tags:
        - requests
  /invites/{id}:
    post:
      summary: Complete invite
      description: >
        Completes the invite and removes the request. The request is also removed if the invite fails with an error other than 400 Invalid data (invalid user data).
      consumes:
        - application/json      
      produces:
        - application/json
      parameters:
        - in: path
          name: id
          required: true
          type: string
          description: Invite id
        - name: data
          in: body
          schema:
            $ref: '#/definitions/CompleteInvite'
      responses:
        200:
          description: Success
          schema:
            $ref: '#/definitions/CompleteInviteResponse'
        400:
          description: Invalid data 
          schema:
            $ref: '#/definitions/Error'
        404:
          description: Not found 
          schema:
            $ref: '#/definitions/Error'
        409:
          description: User not found 
          schema:
            $ref: '#/definitions/Error'
      tags:
        - requests
        
    delete:
      summary: Reject invite
      description: >
        Reject invite. Removes the request, deletes the temporary user account, and adds the email to Do Not Disturb list so that further invites are refused. Note that email is subsequently removed from the DnD list only if the user later on registers to the service.
      consumes:
        - application/json      
      produces:
        - application/json
      parameters:
        - in: path
          name: id
          required: true
          type: string
          description: Invite id
      responses:
        204:
          description: Success
        404:
          description: Not found 
          schema:
            $ref: '#/definitions/Error'
      tags:
        - requests
  /me/login:
    post:
      summary: Start a new session
      description: >
        Start a new session. On success, returns a session token, and session expiration time as numbers of milliseconds elapsed since January 1, 1970 00:00:00 UTC.
      consumes:
        - application/json      
      produces:
        - application/json
      parameters:
        - name: data
          in: body
          schema:
            $ref: '#/definitions/Credentials'
      responses:
        201:
          description: Success
          schema:
            $ref: '#/definitions/Session'
        400:
          description: Invalid credentials 
          schema:
            $ref: '#/definitions/Error'
      tags:
        - me
  /me/renew:
    post:
      summary: Renew the current session
      description: >
        Renew the current session.
      parameters:
        - name: Authorization
          in: header
          description: Session token as 'POSTCARD-TOKEN token="1111-2222-3333-4444"'
          required: true
          type: string
      responses:
        200:
          description: Success
          schema:
            $ref: '#/definitions/Session'
        403:
          description: Invalid session 
          schema:
            $ref: '#/definitions/Error'
      tags:
        - me
  /me/logout:
    post:
      summary: End the current session
      description: >
        End the current session.
      parameters:
        - name: Authorization
          in: header
          description: Session token as 'POSTCARD-TOKEN token="1111-2222-3333-4444"'
          required: true
          type: string
      responses:
        204:
          description: Success
        403:
          description: Invalid session 
          schema:
            $ref: '#/definitions/Error'
      tags:
        - me
  /me:
    get:
      summary: Get current user data
      description: >
        Get current user data.
      parameters:
        - name: Authorization
          in: header
          description: Session token as 'POSTCARD-TOKEN token="1111-2222-3333-4444"'
          required: true
          type: string
      responses:
        200:
          description: Success
          schema:
            $ref: '#/definitions/Me'
        403:
          description: Invalid session 
          schema:
            $ref: '#/definitions/Error'
      tags:
        - me

    delete:
      summary: Delete account
      description: >
        Ends the current session and deletes the account.
      consumes:
        - application/json      
      produces:
        - application/json
      parameters:
        - name: Authorization
          in: header
          description: Session token as 'POSTCARD-TOKEN token="1111-2222-3333-4444"'
          required: true
          type: string
        - name: data
          in: body
          schema:
            $ref: '#/definitions/Password'
      responses:
        204:
          description: Success
        403:
          description: Invalid session, Invalid password
          schema:
            $ref: '#/definitions/Error'
        409:
          description: User not found 
          schema:
            $ref: '#/definitions/Error'
      tags:
        - me
  /me/password:
    put:
      summary: Set user password
      description: >
        Set user password.
      parameters:
        - name: Authorization
          in: header
          description: Session token as 'POSTCARD-TOKEN token="1111-2222-3333-4444"'
          required: true
          type: string
        - name: data
          in: body
          schema:
            $ref: '#/definitions/SetPassword'
      responses:
        204:
          description: Success
        400:
          description: Invalid data 
          schema:
            $ref: '#/definitions/Error'
        403:
          description: Invalid session or password
          schema:
            $ref: '#/definitions/Error'
        409:
          description: User not found
          schema:
            $ref: '#/definitions/Error'
      tags:
        - me
  /me/language:
    put:
      summary: Set user language
      description: >
        Set user language. Currently, only "en" and "fi" are supported.
      parameters:
        - name: Authorization
          in: header
          description: Session token as 'POSTCARD-TOKEN token="1111-2222-3333-4444"'
          required: true
          type: string
        - name: data
          in: body
          schema:
            $ref: '#/definitions/SetLanguage'
      responses:
        204:
          description: Success
        400:
          description: Invalid data 
          schema:
            $ref: '#/definitions/Error'
        403:
          description: Invalid session
          schema:
            $ref: '#/definitions/Error'
        409:
          description: User not found
          schema:
            $ref: '#/definitions/Error'
      tags:
        - me
...
definitions:
  RegistrationAvailability:
    type: object
    properties:
      email:
        type: boolean
  RegistrationRequest:
    type: object
    properties:
      firstName:
        type: string
      lastName:
        type: string
      email:
        type: string
      password:
        type: string
        format: password
      language:
        type: string
    required:
      - firstName
      - lastName
      - email
      - password
      - language
  ExpiringRequestResponse:
    type: object
    properties:
      expires:
        type: integer
        format: int64
        example: 1553028728606
  CompleteRegistrationResponse:
    type: object
    properties:
      firstName:
        type: string
      lastName:
        type: string
      email:
        type: string
    required:
      - firstName
      - lastName
      - email
  PasswordResetRequest:
    type: object
    properties:
      email:
        type: string
    required:
      - email
  CompletePasswordReset:
    type: object
    properties:
      password:
        type: string
        format: password
    required:
      - email
  CompletePasswordResetResponse:
    type: object
    properties:
      firstName:
        type: string
      lastName:
        type: string
      email:
        type: string
    required:
      - firstName
      - lastName
      - email
...
  CompleteInvite:
    type: object
    properties:
      firstName:
        type: string
      lastName:
        type: string
      password:
        type: string
        format: password
      language:
        type: string
    required:
      - firstName
      - lastName
      - passwordName
      - language
...
  CompleteEmailChangeResponse:
    type: object
    properties:
      firstName:
        type: string
      lastName:
        type: string
      email:
        type: string
    required:
      - firstName
      - lastName
      - email
...
  Error:
    type: object
    properties:
      error:
        type: string
...